requirements.txt
Django==4.1.7
djangorestframework==3.14.0
Pillow==10.0.1
Requests==2.31.0
selenium==4.16.0